native: move continue_reading to async
This patch moves the continue_reading to the async file api. The
raw802154 driver needs such function as well, so we move it to a shared
place.
